Update to Retrofit 2.0.0-beta3, OkHttp 3.0.0-RC1.

+++ src/internalDebug/java/com/jakewharton/u2020/ui/debug/DebugView.java
@@ -49,21 +49,20 @@ import com.jakewharton.u2020.ui.misc.EnumAdapter;
 import com.jakewharton.u2020.util.Keyboards;
 import com.jakewharton.u2020.util.Strings;
 import com.squareup.leakcanary.internal.DisplayLeakActivity;
-import com.squareup.okhttp.Cache;
-import com.squareup.okhttp.OkHttpClient;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.StatsSnapshot;
 import java.lang.reflect.Method;
 import java.net.InetSocketAddress;
-import java.net.Proxy;
 import java.util.Locale;
 import java.util.Set;
 import javax.inject.Inject;
 import javax.inject.Named;
+import okhttp3.Cache;
+import okhttp3.OkHttpClient;
 import org.threeten.bp.ZoneId;
 import org.threeten.bp.format.DateTimeFormatter;
 import org.threeten.bp.temporal.TemporalAccessor;
-import retrofit.mock.NetworkBehavior;
+import retrofit2.mock.NetworkBehavior;
 import timber.log.Timber;
 
 import static butterknife.ButterKnife.findById;
@@ -249,11 +248,14 @@ public final class DebugView extends FrameLayout {
         .filter(position -> !networkProxyAddress.isSet() || position != ProxyAdapter.PROXY)
         .subscribe(position -> {
           if (position == ProxyAdapter.NONE) {
-            Timber.d("Clearing network proxy");
-            // TODO: Keep the custom proxy around so you can easily switch back and forth.
-            networkProxyAddress.delete();
-            client.setProxy(null);
-            apiClient.setProxy(null);
+            // Only clear the proxy and restart if one was previously set.
+            if (currentProxyPosition != ProxyAdapter.NONE) {
+              Timber.d("Clearing network proxy");
+              // TODO: Keep the custom proxy around so you can easily switch back and forth.
+              networkProxyAddress.delete();
+              // Force a restart to re-initialize the app without a proxy.
+              ProcessPhoenix.triggerRebirth(getContext());
+            }
           } else if (networkProxyAddress.isSet() && position == ProxyAdapter.PROXY) {
             Timber.d("Ignoring re-selection of network proxy %s", networkProxyAddress.get());
           } else {
@@ -443,21 +445,21 @@ public final class DebugView extends FrameLayout {
   }
 
   private void setupOkHttpCacheSection() {
-    Cache cache = client.getCache(); // Shares the cache with apiClient, so no need to check both.
-    okHttpCacheMaxSizeView.setText(getSizeString(cache.getMaxSize()));
+    Cache cache = client.cache(); // Shares the cache with apiClient, so no need to check both.
+    okHttpCacheMaxSizeView.setText(getSizeString(cache.maxSize()));
 
     refreshOkHttpCacheStats();
   }
 
   private void refreshOkHttpCacheStats() {
-    Cache cache = client.getCache(); // Shares the cache with apiClient, so no need to check both.
-    int writeTotal = cache.getWriteSuccessCount() + cache.getWriteAbortCount();
-    int percentage = (int) ((1f * cache.getWriteAbortCount() / writeTotal) * 100);
+    Cache cache = client.cache(); // Shares the cache with apiClient, so no need to check both.
+    int writeTotal = cache.writeSuccessCount() + cache.writeAbortCount();
+    int percentage = (int) ((1f * cache.writeAbortCount() / writeTotal) * 100);
     okHttpCacheWriteErrorView.setText(
-        cache.getWriteAbortCount() + " / " + writeTotal + " (" + percentage + "%)");
-    okHttpCacheRequestCountView.setText(String.valueOf(cache.getRequestCount()));
-    okHttpCacheNetworkCountView.setText(String.valueOf(cache.getNetworkCount()));
-    okHttpCacheHitCountView.setText(String.valueOf(cache.getHitCount()));
+        cache.writeAbortCount() + " / " + writeTotal + " (" + percentage + "%)");
+    okHttpCacheRequestCountView.setText(String.valueOf(cache.requestCount()));
+    okHttpCacheNetworkCountView.setText(String.valueOf(cache.networkCount()));
+    okHttpCacheHitCountView.setText(String.valueOf(cache.hitCount()));
   }
 
   private void applyAnimationSpeed(int multiplier) {
@@ -526,13 +528,9 @@ public final class DebugView extends FrameLayout {
           String in = hostView.getText().toString();
           InetSocketAddress address = InetSocketAddressPreferenceAdapter.parse(in);
           if (address != null) {
-            networkProxyAddress.set(address); // Persist across restarts.
-            proxyAdapter.notifyDataSetChanged(); // Tell the spinner to update.
-            networkProxyView.setSelection(ProxyAdapter.PROXY); // And show the proxy.
-
-            Proxy proxy = InetSocketAddressPreferenceAdapter.createProxy(address);
-            client.setProxy(proxy);
-            apiClient.setProxy(proxy);
+            networkProxyAddress.set(address);
+            // Force a restart to re-initialize the app with the new proxy.
+            ProcessPhoenix.triggerRebirth(getContext());
           } else {
             networkProxyView.setSelection(originalSelection);
           }
